schema: move ConnectionInfo to schema
